Moved versions method to workspace_fs.py and adapted code
